backend/requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
django-cors-headers
djangorestframework-simplejwt
